Silence warning

+++ Editor/Steamworks.NET/RedistInstall.cs
@@ -36,11 +36,6 @@ public class RedistInstall {
 		string strSource = Path.Combine(Path.Combine(strCWD, path), filename);
 		string strDest = Path.Combine(strCWD, filename);
 
-		if (!File.Exists(strSource)) {
-			Debug.LogWarning(string.Format("[Steamworks.NET] Could not copy {0} into the project root. {0} could not be found in '{1}'. Place {0} from the Steamworks SDK in the project root manually.", filename, Path.Combine(strCWD, path)));
-			return;
-		}
-
 		if (File.Exists(strDest)) {
 			if (!bCheckDifference)
 				return;
@@ -57,6 +52,11 @@ public class RedistInstall {
 		}
 		else {
 			Debug.Log(string.Format("[Steamworks.NET] {0} is not present in the project root. Copying...", filename));
+		}
+
+		if (!File.Exists(strSource)) {
+			Debug.LogWarning(string.Format("[Steamworks.NET] Could not copy {0} into the project root. {0} could not be found in '{1}'. Place {0} from the Steamworks SDK in the project root manually.", filename, Path.Combine(strCWD, path)));
+			return;
 		}
 
 		File.Copy(strSource, strDest, true);
